fix: refactored execute.c
