+++ includes/minishell.h
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-# ifndef MINISHELL_H
+#ifndef MINISHELL_H
 # define MINISHELL_H
 
 # define INT_MAX 2147483647
@@ -74,7 +74,6 @@ typedef struct s_hash_table
 {
 	t_ht_item		**item;
 	int				size;
-	int				count; //o que que é count??
 }	t_hash_table;
 
 typedef struct s_shell
@@ -141,6 +140,8 @@ char			*put_quotes(t_ht_item *new_env);
 char			**env_with_quotes(void);
 void			print_export_env(char **array, int fd);
 void			error_export(char **cmd, int i);
+int				is_builtins(char **cmd);
+void			builtins(char **cmd, char **old_cmd, char **n_env);
 
 /*
 ** HASH TABLE
@@ -191,6 +192,8 @@ char			*get_path_str(void);
 char			**get_paths(void);
 char			*do_prompt(void);
 int				is_path(char **cmd, char **n_env);
+char			**create_command_for_exec(char **cmd, char **paths);
+void			do_exec(char **cmd, char **n_env);
 
 /*
 ** UTILS
